+++ commons/src/main/java/com/karuslabs/commons/command/tree/nodes/Literal.java
@@ -23,7 +23,7 @@
  */
 package com.karuslabs.commons.command.tree.nodes;
 
-import com.karuslabs.commons.command.Commands;
+import com.karuslabs.commons.command.*;
 
 import com.mojang.brigadier.*;
 import com.mojang.brigadier.tree.*;
@@ -70,6 +70,11 @@ public class Literal<T> extends LiteralCommandNode<T> implements Aliasable<T>, M
     public Literal(String name, Command<T> command, Predicate<T> requirement) {
         this(name, command, requirement, null, null, false);
     }
+    
+    public Literal(String name, Execution<T> execution, Predicate<T> requirement) {
+        this(name, execution, requirement, null, null, false);
+    }
+    
     
     public Literal(String name, Command<T> command, Predicate<T> requirement, @Nullable CommandNode<T> destination, RedirectModifier<T> modifier, boolean fork) {
         this(name, new ArrayList<>(0), false, command, requirement, destination, modifier, fork);